Some minor fixes; enabled data checksums for Debian installations.
Enabling data checksums is only possible when creating a cluster. As the Debian package always initializes a cluster with defaults, that cluster is dropped and a new one is created. This is only done when there did not exist a cluster directory before installing the packages (i.e. the fact is set at the beginning of install_apt.yml).