Added fencing support (primarily for HAProxy) - see README.md for details
repmgr does not provide real support for node fencing. There is an example in the Github repo which fences off a failed primary by changing the PgBouncer config correspondingly. Unfortunately that does not work for PgBouncer instances that are not available during promoting the new primary.

This commit contains a 'hands-on' approach to solve that problem. It would be way better if repmgr/repmgrd could simply return the hostname of the current primary, but as long as that is not available we need at least some kind of solution - otherwise automatic failover would not make any sense at all.